% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{analyze_outcome}
\alias{analyze_outcome}
\title{Analyze Pattern-Outcome Relationships}
\usage{
analyze_outcome(
  data,
  cols = tidyselect::everything(),
  id = NULL,
  group = "last_obs",
  n = 10,
  type = "presence",
  priority = "separation",
  formula = ~1,
  re_formula,
  ...
)
}
\arguments{
\item{data}{[\code{data.frame}]\cr Sequence data in wide format
(rows are sequences, columns are time points). The input should be
coercible to a \code{data.frame} object.}

\item{cols}{[\code{expression}]\cr A tidy selection of columns that should
be considered as sequence data. By default, all columns are used.}

\item{id}{[\code{expression}]\cr An optional tidy selection of columns that
define the hierarchy of the data. If provided, random effects
can be specified via \code{re_formula}. The default value \code{NULL} corresponds
to a fixed-effects model.}

\item{group}{[\code{character(1)}]\cr Outcome group specification. The
option \code{"last_obs"} (default) assumes that the last non-missing observation
of each sequence specifies the outcome groups. This can also be a column
name of \code{data} that contains the outcome group variable information.}

\item{n}{[\code{integer(1)}]\cr Maximum number of patterns to include in the
model as covariates. The default is \code{10}.}

\item{type}{[\code{character(1)}]\cr Specifies how the patterns are included
as predictors. The option \code{"count"} uses the number of times a specific
pattern occurs in a sequence, whereas option \code{"presence"} (the default)
defines a binary predictors that attains the value \code{1} if the pattern is
present in the sequence and \code{0} otherwise.}

\item{priority}{[\code{character(1)}]\cr Priority of pattern inclusion in the
model. The available options are:
\itemize{
\item \code{"separation"}: Best separation of the \code{outcome} classes. For
\code{type = "count"}, the chi-square statistic is used. For
\code{type = "presence"}, the proportion of 1-1 and 0-0 matches between
the pattern and the \code{outcome} is used.
\item \code{"frequency"}: Most frequent patterns.
\item \code{"support"}: Most common patterns.
}}

\item{formula}{[\code{formula}]\cr Formula specification for the fixed effects
of the non-pattern covariates. The default is \code{~ 1}, adding no covariates.}

\item{re_formula}{[\code{formula}]\cr Formula specification of the random
effects when \code{id} is provided. By default, a random intercept is added
for each variable.}

\item{...}{Additional arguments passed to \code{\link[=discover_patterns]{discover_patterns()}} and the
model-fitting function (\code{\link[stats:glm]{stats::glm()}} or \code{\link[lme4:glmer]{lme4::glmer()}}).}
}
\value{
Either a \code{glm} or a \code{glmerMod} object depending on whether
random effects were included.
}
\description{
Fit a (mixed) logistic regression model to the data using sequence patterns
as predictors. The patterns are first determined using \code{\link[=discover_patterns]{discover_patterns()}}
and used as predictors as specified by the user (either count or presence).
The user can further select the maximum number of patterns to use and how
to prioritize the patterns (separation, frequency, proportion of sequences,
etc.). Additional covariates and random effects can also be included in the
model. The logistic model is fitted using \code{\link[stats:glm]{stats::glm()}} or by
\code{\link[lme4:glmer]{lme4::glmer()}} in the case of a mixed model.
}
\examples{
# TODO
1 + 1

}
