% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codyna-package.R
\docType{package}
\name{codyna-package}
\alias{codyna}
\alias{codyna-package}
\title{The \code{codyna} Package.}
\description{
Performs analysis of complex dynamic systems with a focus on the
temporal unfolding of patterns, changes, and state transitions in
behavioral data. The package supports both time series and sequence data
and provides tools for the analysis and visualization of complexity,
pattern identification, trends, regimes, sequence typology as well as
early warning signals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/santikka/codyna/}
  \item Report bugs at \url{https://github.com/santikka/codyna/issues/}
}

}
\author{
Santtu Tikka and Mohammed Saqr
}
