% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codyna-package.R
\docType{data}
\name{ema}
\alias{ema}
\title{Ecological Momentary Assessment (EMA) Data}
\format{
A \code{data.frame} object.
}
\source{
\url{https://github.com/heinonmatti/complexity-behchange}
}
\usage{
ema
}
\description{
Example data for complex adaptive systems perspective to behavior change
research. The dataset consists of 20 individuals with 9 self-report
variables (and time of response) each. For more information on the data,
please see
\url{https://heinonmatti.github.io/complexity-behchange/dataset-info.html}
}
\keyword{datasets}
