% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ews}
\alias{plot.ews}
\title{Plot EWS Results}
\usage{
\method{plot}{ews}(x, ...)
}
\arguments{
\item{x}{[\code{ews}]\cr
Output of \code{\link[=detect_warnings]{detect_warnings()}}.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot EWS Results
}
\examples{
set.seed(123)
ts_data <- stats::arima.sim(list(order = c(1, 1, 0), ar = 0.6), n = 200)
ews_roll <- detect_warnings(ts_data)
plot(ews_roll)

}
