% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.patterns}
\alias{plot.patterns}
\title{Plot Discovered Patterns}
\usage{
\method{plot}{patterns}(x, n = 10, ...)
}
\arguments{
\item{x}{[\code{patterns}]\cr Output of \code{\link[=discover_patterns]{discover_patterns()}}.}

\item{n}{[\code{integer(1)}]\cr Maximum number of patterns
to include in the plot.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} object.
ngrams <- discover_patterns(engagement, type = "ngram")
plot(ngrams)
}
\description{
Plot Discovered Patterns
}
