% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.changepoint}
\alias{print.changepoint}
\title{Print Changepoint Detection Results}
\usage{
\method{print}{changepoint}(x, ...)
}
\arguments{
\item{x}{[\code{changepoint}]\cr A \code{changepoint} object.}

\item{...}{Additional arguments passed to the tibble print method.}
}
\value{
\code{x}, invisibly.
}
\description{
Print Changepoint Detection Results
}
