% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ews}
\alias{print.ews}
\title{Print EWS Detection Results}
\usage{
\method{print}{ews}(x, ...)
}
\arguments{
\item{x}{[\code{ews}]\cr
EWS detection result from \code{\link[=detect_warnings]{detect_warnings()}}.}

\item{...}{Additional arguments passed to the generic print method.}
}
\value{
\code{x} (invisibly).
}
\description{
Print EWS Detection Results
}
\examples{
set.seed(123)
ts_data <- stats::arima.sim(list(order = c(1, 1, 0), ar = 0.6), n = 200)
ews <- detect_warnings(ts_data)
print(ews)

}
