% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.multi_ews}
\alias{print.multi_ews}
\title{Print a Multivariate EWS Object}
\usage{
\method{print}{multi_ews}(x, ...)
}
\arguments{
\item{x}{[\code{multi_ews}]\cr A \code{multi_ews} object.}

\item{...}{Additional arguments passed to the tibble print method.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{"multi_ews"} objects.
Delegates to tibble's default print.
}
