% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.resilience}
\alias{print.resilience}
\title{Print a Resilience Object}
\usage{
\method{print}{resilience}(x, ...)
}
\arguments{
\item{x}{[\code{resilience}]\cr
A \code{resilience} object.}

\item{...}{Additional arguments passed to the tibble print method.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{"resilience"} objects. Dispatches
to the tibble print method, displaying the time series data and computed
resilience metrics.
}
