% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.sensitivity_ews}
\alias{print.sensitivity_ews}
\title{Print a Sensitivity Object}
\usage{
\method{print}{sensitivity_ews}(x, ...)
}
\arguments{
\item{x}{[\code{sensitivity_ews}]\cr
A \code{sensitivity_ews} object.}

\item{...}{Additional arguments passed to the tibble print method.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{"sensitivity_ews"} objects.
Dispatches to the tibble print method.
}
