% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.spectral}
\alias{print.spectral}
\title{Print a Spectral EWS Analysis Results}
\usage{
\method{print}{spectral}(x, ...)
}
\arguments{
\item{x}{[\code{spectral}]\cr
A \code{spectral} object.}

\item{...}{Additional arguments passed to the tibble print method.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{"spectral"} objects. Dispatches to the tibble print method.
}
