% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.changepoint}
\alias{print.summary.changepoint}
\title{Print a Summary of Changepoint Detection Results}
\usage{
\method{print}{summary.changepoint}(x, ...)
}
\arguments{
\item{x}{[\code{summary.changepoint}]\cr A \code{summary.changepoint} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Summary of Changepoint Detection Results
}
