% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.hurst_ews}
\alias{print.summary.hurst_ews}
\title{Print Hurst Early Warning Signal Analysis Summary}
\usage{
\method{print}{summary.hurst_ews}(x, ...)
}
\arguments{
\item{x}{[\code{summary.hurst_ews}]\cr A \code{summary.hurst_ews} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print Hurst Early Warning Signal Analysis Summary
}
