% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.multi_ews}
\alias{print.summary.multi_ews}
\title{Print a Summary of a  Multivariate EWS Object}
\usage{
\method{print}{summary.multi_ews}(x, ...)
}
\arguments{
\item{x}{[\code{summary.multi_ews}]\cr A \code{summary.multi_ews} object.}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Summary of a  Multivariate EWS Object
}
