% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.sensitivity_ews}
\alias{print.summary.sensitivity_ews}
\title{Print a EWS Sensitivity Analysis Summary}
\usage{
\method{print}{summary.sensitivity_ews}(x, ...)
}
\arguments{
\item{x}{[\code{summary.sensitivity_ews}]\cr A \code{summary.sensitivity_ews} object.}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a EWS Sensitivity Analysis Summary
}
