% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.spectral}
\alias{print.summary.spectral}
\title{Print a Summary of Spectral EWS Analysis Results}
\usage{
\method{print}{summary.spectral}(x, ...)
}
\arguments{
\item{x}{[\code{summary.spectral}]\cr A \code{summary.spectral} object.}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Summary of Spectral EWS Analysis Results
}
