% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.surrogate_test}
\alias{print.summary.surrogate_test}
\title{Print a Summary of Surrogate Test Results}
\usage{
\method{print}{summary.surrogate_test}(x, ...)
}
\arguments{
\item{x}{[\code{summary.surrotate_test}]\cr A \code{summary.surrogate_test} object.}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Summary of Surrogate Test Results
}
