% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.trend}
\alias{print.summary.trend}
\title{Print a Trend Classification Summary}
\usage{
\method{print}{summary.trend}(x, ...)
}
\arguments{
\item{x}{[\code{summary.trend}]\cr A \code{summary.trend} object.}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Trend Classification Summary
}
