% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.trend}
\alias{print.trend}
\title{Print a Trend Object}
\usage{
\method{print}{trend}(x, ...)
}
\arguments{
\item{x}{[\code{trend}]\cr
A \code{trend} object.}

\item{...}{Additional arguments passed to the tibble print method.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{"trend"} objects. Dispatches to the tibble print method.
}
