% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{sequence_indices}
\alias{sequence_indices}
\title{Compute Sequence Indices for Sequence Data}
\usage{
sequence_indices(data, cols, favorable, omega = 1, ...)
}
\arguments{
\item{data}{[\code{data.frame}, \code{matrix}, \code{stslist}]\cr
Sequence data in wide format (rows are sequences, columns are time points).}

\item{cols}{[\code{expression}]\cr A tidy selection of columns that should
be considered as sequence data. By default, all columns are used.}

\item{favorable}{A \code{character} vector of state names that should be
considered as favorable states.}

\item{omega}{A \code{numeric} value for the omega parameter used to compute
the integrative potential.}

\item{...}{Not used.}
}
\value{
A \code{tibble} containing the index values.
}
\description{
Compute Sequence Indices for Sequence Data
}
\examples{
# TODO
x <- rnorm(1)

}
