% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.patterns}
\alias{plot.patterns}
\title{Plot Discovered Patterns}
\usage{
\method{plot}{patterns}(x, n = 10, prop = TRUE, group, global = TRUE, ...)
}
\arguments{
\item{x}{[\code{patterns}]\cr Output of \code{\link[=discover_patterns]{discover_patterns()}}.}

\item{n}{[\code{integer(1)}]\cr Maximum number of patterns
to include in the plot.}

\item{prop}{[\code{logical(1)}]\cr Should group-specific count proportions be
displayed in the plot? The default is \code{TRUE}. Ignored if the original
sequences were not grouped.}

\item{group}{[\code{character(1)}]\cr Name of the group to draw the plot for.
If not provided (the default), shows the counts and proportions by
group for each pattern. If provided, only the proportions of the
specific group are drawn by pattern. Ignored if the original sequences
were not grouped.}

\item{global}{[\code{logical(1)}]\cr Should a line be added showing the global
proportion when \code{group} is provided? (default: \code{TRUE}). Also colors
the patterns according to whether the proportion is above or below
the global value.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} object.
ngrams <- discover_patterns(engagement, type = "ngram")
plot(ngrams)
}
\description{
Plot Discovered Patterns
}
